modularized our transforms, added a __mock__ folder with pictures, and make a test file in __tests__
